voice recognition separated and coroutines
