+++ src/main/kotlin/Main.kt
@@ -11,6 +11,7 @@ import kotlinx.coroutines.flow.takeWhile
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.runBlocking
 import org.example.chatmodel.ChatModel
+import org.example.voice.recognition.VoiceRecognition
 import org.vosk.LibVosk
 import org.vosk.LogLevel
 import org.vosk.Model
@@ -18,18 +19,7 @@ import org.vosk.Recognizer
 import javax.sound.sampled.*
 import kotlin.concurrent.thread
 
-fun buildAudioInputStream(): TargetDataLine? {
-    val audioFormat = AudioFormat(16000f, 16, 1, true, false)
-    val info = DataLine.Info(TargetDataLine::class.java, audioFormat)
-    if (!AudioSystem.isLineSupported(info)) {
-        println("Line is not supported!")
-        return null
-    }
-    val microphone = AudioSystem.getLine(info) as TargetDataLine
-    microphone.open(audioFormat)
-    microphone.start()
-    return microphone
-}
+
 
 fun messageOllama(model: OllamaStreamingChatModel, message: String, speak: Espeak) {
 
@@ -45,39 +35,12 @@ fun main(): Unit = runBlocking {
 
     val speak = Espeak(voice)
 
-    LibVosk.setLogLevel(LogLevel.DEBUG)
-    val vosk_model = Model("vosk_models/vosk-model-small-en-us-0.15")
-    val recognizer = Recognizer(vosk_model, 16000f)
-
-    println("vosk initialized")
-
-    var nbytes: Int
-
-
-    val mic = buildAudioInputStream()// AudioSystem.getAudioInputStream(BufferedInputStream(FileInputStream("test.wav")))
-    launch {
-        val bytes = ByteArray(4096)
-
-        var last = ""
-        while (true) {
-            nbytes = mic?.read(bytes, 0, bytes.size) ?: break
-            if (nbytes < 0)
-                break
-
-            if (nbytes > 0)
-            if (recognizer.acceptWaveForm(bytes, nbytes)) {
-                if (last != recognizer.result) {
-                    val message = last.split("\"")[3]
-                    if (message.isNotEmpty())
-                        ChatModel.message(message).takeWhile { it != null }.collect { token ->
-                            println("lel $token")
-                        }
-                    last = recognizer.result
-                }
-            } else {
-                last = recognizer.partialResult
-            }
-        }
+    /*
+    ChatModel.message(message).takeWhile { it != null }.collect { token ->
+        println("lel $token")
     }
+    */
+
+    VoiceRecognition.listen().collect(::println)
 
 }
